create class dynamodb

+++ poc-insurance-example/market/src/main/java/br/com/insurance/market/domain/service/CoverTax.java
@@ -2,5 +2,5 @@ package br.com.insurance.market.domain.service;
 
 public interface CoverTax {
 
-    Integer getCoverTax(String productCode);
+    double getCoverTax(String productCode);
 }
